New FeatherBB middleware architecture
