+++ Slim/Extras/Middleware/FeatherBBLoader.php
@@ -1,0 +1,258 @@
+<?php
+/**
+ *
+ * Copyright (C) 2015 FeatherBB
+ * based on code by (C) 2008-2012 FluxBB
+ * and Rickard Andersson (C) 2002-2008 PunBB
+ * License: http://www.gnu.org/licenses/gpl.html GPL version 2 or higher
+ *
+ * $app = new \Slim\Slim();
+ * $app->add(new \Slim\Extras\Middleware\FeatherBBLoader(array $config));
+ *
+ */
+
+namespace Slim\Extras\Middleware;
+use DB;
+
+class FeatherBBLoader extends \Slim\Middleware
+{
+    protected $forum_env,
+              $forum_settings;
+    protected $headers = array(
+        'Cache-Control' => 'no-cache, no-store, must-revalidate',
+        'Pragma' => 'no-cache',
+        'Content-type' => 'text/html',
+        'X-Frame-Options' => 'deny');
+
+	public function __construct(array $data)
+	{
+        // Define some core variables
+        $this->forum_env['FEATHER_ROOT'] = realpath(dirname(__FILE__). '../../../');
+        $this->forum_env['FORUM_CACHE_DIR'] = is_writable($this->forum_env['FEATHER_ROOT'].$data['cache_dir']) ? realpath($this->forum_env['FEATHER_ROOT'].$data['cache_dir']).'/' : null;
+        $this->forum_env['FORUM_CONFIG_FILE'] = is_file($this->forum_env['FEATHER_ROOT'].$data['config_file']) ? realpath($this->forum_env['FEATHER_ROOT'].$data['config_file']) : null;
+        // Populate forum_env
+        $this->forum_env = array_merge(self::load_default_forum_env(), $this->forum_env);
+        $this->env_to_globals($this->forum_env); // Legacy
+
+        // Load files
+        require $this->forum_env['FEATHER_ROOT'].'include/utf8/utf8.php'; // ?
+        require $this->forum_env['FEATHER_ROOT'].'include/functions.php';
+        require $this->forum_env['FEATHER_ROOT'].'include/pomo/MO.php';
+        require $this->forum_env['FEATHER_ROOT'].'include/l10n.php';
+        require $this->forum_env['FEATHER_ROOT'].'include/idiorm.php';
+
+        // Force POSIX locale (to prevent functions such as strtolower() from messing up UTF-8 strings)
+        setlocale(LC_CTYPE, 'C');
+        // Reset opcache while debugging
+        opcache_reset();
+	}
+
+    public static function load_default_forum_env()
+    {
+        return array(
+                'FEATHER' => true, // Legacy
+                'FEATHER_ROOT' => '',
+                'FORUM_CONFIG_FILE' => 'include/config.php',
+                'FORUM_CACHE_DIR' => 'cache/',
+                'FORUM_VERSION' => '1.0.0',
+                'FORUM_NAME' => 'FeatherBB',
+                'FORUM_DB_REVISION' => 21,
+                'FORUM_SI_REVISION' => 2,
+                'FORUM_PARSER_REVISION' => 2,
+                'FEATHER_UNVERIFIED' => 0,
+                'FEATHER_ADMIN' => 1,
+                'FEATHER_MOD' => 2,
+                'FEATHER_GUEST' => 3,
+                'FEATHER_MEMBER' => 4,
+                'FEATHER_MAX_POSTSIZE' => 32768,
+                'FEATHER_SEARCH_MIN_WORD' => 3,
+                'FEATHER_SEARCH_MAX_WORD' => 20,
+                'FORUM_MAX_COOKIE_SIZE' => 4048,
+                'FEATHER_DEBUG' => 1,
+                'FEATHER_SHOW_QUERIES' => 1,
+                'FEATHER_CACHE_QUERIES' => 0,
+                );
+    }
+
+    public static function load_default_forum_settings()
+    {
+        return array(
+                // Database
+                'db_type' => 'mysqli',
+                'db_host' => '',
+                'db_name' => '',
+                'db_user' => '',
+                'db_pass' => '',
+                'db_prefix' => '',
+                // Cookies
+                'cookie_name' => 'feather_cookie',
+                'cookie_seed' => 'changeme', // MUST BE CHANGED !!!
+                // Debug
+                'debug' => false,
+                );
+    }
+
+    public function init_db()
+    {
+        switch ($this->forum_settings['db_type']) {
+            case 'mysql':
+            case 'mysqli':
+            case 'mysql_innodb':
+            case 'mysqli_innodb':
+                DB::configure('mysql:host='.$this->forum_settings['db_host'].';dbname='.$this->forum_settings['db_name']);
+                DB::configure('driver_options', array(\PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8'));
+                break;
+            case 'sqlite';
+            case 'sqlite3';
+                DB::configure('sqlite:./'.$this->forum_settings['db_name']);
+                break;
+            case 'pgsql':
+                \DB::configure('pgsql:host='.$this->forum_settings['db_host'].'dbname='.$this->forum_settings['db_name']);
+                break;
+        }
+        DB::configure('username', $this->forum_settings['db_user']);
+        DB::configure('password', $this->forum_settings['db_pass']);
+        if ($this->forum_env['FEATHER_SHOW_QUERIES'] == 1) {
+            DB::configure('logging', true);
+        }
+        if ($this->forum_env['FEATHER_CACHE_QUERIES'] == 1) {
+            DB::configure('caching', true);
+        }
+        DB::configure('id_column_overrides', array(
+            $this->forum_settings['db_prefix'].'groups' => 'g_id',
+        ));
+    }
+
+    // Getters / setters for Slim container (avoid magic get error)
+
+    public function set_forum_env($key, $value = null)
+    {
+        $tmp = (!is_array($key) && !is_null($value)) ? array($key, $value) : $key;
+        foreach ($tmp as $key => $value) {
+            $this->app->container->get('forum_env')[$key] = $value;
+        }
+
+    }
+
+    public function set_forum_settings($key, $value = null)
+    {
+        $tmp = (!is_array($key) && !is_null($value)) ? array($key, $value) : $key;
+        foreach ($tmp as $key => $value) {
+            $this->app->container->get('forum_settings')[$key] = $value;
+        }
+    }
+
+    // Legacy function, to ensure backward compatibility with globals
+    public function env_to_globals(array $vars)
+    {
+        foreach ($vars as $key => $value) {
+            define($key, $value);
+        }
+    }
+
+    public function hydrate($name, array $data)
+    {
+        $this->app->container[$name] = $data;
+    }
+
+    // Headers
+
+    public function set_headers()
+    {
+        foreach ($this->headers as $label => $value) {
+            $this->app->response->headers->set($label, $value);
+        }
+        $this->app->response()->headers()->set('X-Powered-By', $this->forum_env['FORUM_NAME']);
+        $this->app->expires(0);
+    }
+
+    public function maintenance_message()
+    {
+        // Deal with newlines, tabs and multiple spaces
+        $pattern = array("\t", '  ', '  ');
+        $replace = array('&#160; &#160; ', '&#160; ', ' &#160;');
+        $message = str_replace($pattern, $replace, $this->forum_settings['o_maintenance_message']);
+
+        $page_title = array(feather_escape($this->forum_settings['o_board_title']), __('Maintenance'));
+
+        define('FEATHER_ACTIVE_PAGE', 'index');
+
+        $this->app->config('templates.path', (is_dir('style/'.$this->app->user->style.'/view')) ? FEATHER_ROOT.'style/'.$this->app->user->style.'/view' : FEATHER_ROOT.'view');
+
+        $header = new \controller\header();
+
+        $header->setTitle($page_title)->display();
+
+        $this->app->render('message.php', array(
+                'message'    =>    $message,
+                'no_back_link'    =>    '',
+            )
+        );
+
+        $footer = new \controller\footer();
+
+        $footer->dontStop();
+
+        $footer->display();
+    }
+
+    public function call()
+    {
+        global $forum_time_formats, $forum_date_formats, $feather_config; // Legacy
+
+        // Set headers
+        $this->set_headers();
+
+        // Record start time
+        $this->app->start = get_microtime();
+
+        // Populate FeatherBB Slim object with forum_env vars
+        $this->hydrate('forum_env', $this->forum_env);
+
+        if ((isset($this->app->environment['HTTP_X_MOZ'])) && ($this->app->environment['HTTP_X_MOZ'] == 'prefetch')) { // Block prefetch requests
+            $this->set_headers();
+            return $this->app->response->setStatus(403); // Send forbidden header
+        }
+
+        if (is_null($this->forum_env['FORUM_CONFIG_FILE'])) {
+            echo 'install !!!!';
+            return;
+        }
+
+        $config_file = json_decode(file_get_contents($this->forum_env['FORUM_CONFIG_FILE']), true);
+        if (!is_null($config_file)) {
+            $this->forum_settings = array_merge(self::load_default_forum_settings(), $config_file);
+        } else {
+            // Erreur
+            echo 'Wrong config format';
+        }
+
+        // Init DB
+        $this->init_db();
+
+        // Get forum settings from DB/cache and load it into forum_settings array
+        if (file_exists($this->forum_env['FORUM_CACHE_DIR'].'cache_config.php')) {
+            include $this->forum_env['FORUM_CACHE_DIR'].'cache_config.php';
+        } else {
+            require $this->forum_env['FEATHER_ROOT'].'include/cache.php';
+            generate_config_cache();
+            require $this->forum_env['FORUM_CACHE_DIR'].'cache_config.php';
+        }
+
+        // Finalize forum_settings array
+        $this->forum_settings = array_merge($feather_config, $this->forum_settings);
+
+        // Populate FeatherBB Slim object with forum_settings vars
+        $this->hydrate('forum_settings', $this->forum_settings);
+        $this->app->config = $this->forum_settings; // Legacy
+        var_dump($this->app->config);
+        extract($this->forum_settings); // Legacy
+
+        // Define time formats
+        $forum_time_formats = array($this->forum_settings['o_time_format'], 'H:i:s', 'H:i', 'g:i:s a', 'g:i a');
+        $forum_date_formats = array($this->forum_settings['o_date_format'], 'Y-m-d', 'Y-d-m', 'd-m-Y', 'm-d-Y', 'M j Y', 'jS M Y');
+
+        // Call FeatherBBAuth middleware
+        $this->next->call();
+    }
+}
